fix test_lightning_trainer not found error
