+++ python/ray/train/tests/test_lightning_trainer.py
@@ -11,79 +11,9 @@ from ray.train.tests.dummy_preprocessor import DummyPreprocessor
 from ray.air.util.data_batch_conversion import convert_batch_type_to_pandas
 import pytest
 
-
-class LinearModule(pl.LightningModule):
-    def __init__(self, input_dim, output_dim) -> None:
-        super().__init__()
-        self.linear = nn.Linear(input_dim, output_dim)
-
-    def forward(self, input):
-        return self.linear(input)
-
-    def training_step(self, batch):
-        output = self.forward(batch)
-        loss = torch.sum(output)
-        self.log("loss", loss)
-        return loss
-    
-    def validation_step(self, val_batch, batch_idx):
-        loss = self.forward(val_batch)
-        return {"val_loss": loss}
-    
-    def validation_epoch_end(self, outputs) -> None:
-        avg_loss = torch.stack([x["val_loss"] for x in outputs]).mean()
-        self.log("val_loss", avg_loss)
-
-    def predict_step(self, batch, batch_idx):
-        return self.forward(batch)
-
-    def configure_optimizers(self):
-        return torch.optim.SGD(self.parameters(), lr=0.1)
-
-
-class DoubleLinearModule(pl.LightningModule):
-    def __init__(self, input_dim_1, input_dim_2, output_dim) -> None:
-        super().__init__()
-        self.linear_1 = nn.Linear(input_dim_1, output_dim)
-        self.linear_2 = nn.Linear(input_dim_2, output_dim)
-
-    def forward(self, batch):
-        input_1 = batch["input_1"]
-        input_2 = batch["input_2"]
-        return self.linear_1(input_1) + self.linear_2(input_2)
-
-    def training_step(self, batch):
-        output = self.forward(batch)
-        loss = torch.sum(output)
-        self.log("loss", loss)
-        return loss
-    
-    def validation_step(self, val_batch, batch_idx):
-        loss = self.forward(val_batch)
-        return {"val_loss": loss}
-    
-    def validation_epoch_end(self, outputs) -> None:
-        avg_loss = torch.stack([x["val_loss"] for x in outputs]).mean()
-        self.log("val_loss", avg_loss)
-
-    def predict_step(self, batch, batch_idx):
-        return self.forward(batch)
-
-    def configure_optimizers(self):
-        return torch.optim.SGD(self.parameters(), lr=0.1)
-
-class DummyDataModule(pl.LightningDataModule):
-    def __init__(self, batch_size: int = 8, dataset_size: int = 256) -> None:
-        super().__init__()
-        self.batch_size = batch_size
-        self.train_data = np.random.rand(dataset_size, 32).astype(np.float32)
-        self.val_data = np.random.rand(dataset_size, 32).astype(np.float32)
-    
-    def train_dataloader(self):
-        return DataLoader(self.train_data, batch_size=self.batch_size)
-    
-    def val_dataloader(self):
-        return DataLoader(self.val_data, batch_size=self.batch_size)
+from ray.train.tests.test_lightning_utils import (
+    LinearModule, DoubleLinearModule, DummyDataModule
+)
 
 
 @pytest.mark.parametrize("accelerator", ["cpu", "gpu"])
